+++ components/airtable_oauth/actions/list-records-in-view/list-records-in-view.mjs
@@ -4,11 +4,16 @@ import commonList from "../common/common-list.mjs";
 export default {
   key: "airtable_oauth-list-records-in-view",
   name: "List Records in View",
-  description: "Retrieve records in a view with automatic pagination. Optionally sort and filter results. Only available for Enterprise accounts.",
+  description: "Retrieve records from a view, optionally sorting and filtering results. [See the documentation](https://airtable.com/developers/web/api/list-views)",
   type: "action",
-  version: "0.0.7",
+  version: "0.0.8",
   ...commonList,
   props: {
+    accountTierAlert: {
+      type: "alert",
+      alertType: "info",
+      content: "Note: views are only available for Airtable Enterprise accounts. [See the documentation](https://airtable.com/developers/web/api/list-views) for more information.",
+    },
     ...common.props,
     viewId: {
       propDefinition: [