+++ components/airtable_oauth/sources/new-records/new-records.mjs
@@ -1,78 +1,19 @@
-import base from "../common/common.mjs";
-import moment from "moment";
+import common from "../common/common-webhook-record.mjs";
 
 export default {
-  ...base,
-  name: "New Records",
+  ...common,
+  name: "New Record(s) Created (Instant)",
   description: "Emit new event for each new record in a table",
   key: "airtable_oauth-new-records",
-  version: "0.0.7",
+  version: "1.0.0",
   type: "source",
-  props: {
-    ...base.props,
-    tableId: {
-      propDefinition: [
-        base.props.airtable,
-        "tableId",
-        ({ baseId }) => ({
-          baseId,
-        }),
-      ],
-      description: "The table ID to watch for changes.",
+  dedupe: "unique",
+  methods: {
+    ...common.methods,
+    getChangeTypes() {
+      return [
+        "add",
+      ];
     },
-    returnFieldsByFieldId: {
-      propDefinition: [
-        base.props.airtable,
-        "returnFieldsByFieldId",
-      ],
-    },
-  },
-  async run() {
-    const {
-      baseId,
-      tableId,
-      viewId,
-    } = this;
-
-    const lastTimestamp = this._getLastTimestamp();
-    const params = {
-      filterByFormula: `CREATED_TIME() > "${lastTimestamp}"`,
-      returnFieldsByFieldId: this.returnFieldsByFieldId || false,
-    };
-
-    const records = await this.airtable.listRecords({
-      baseId,
-      tableId,
-      params,
-    });
-
-    if (!records.length) {
-      console.log("No new records.");
-      return;
-    }
-
-    const metadata = {
-      baseId,
-      tableId,
-      viewId,
-    };
-
-    let maxTimestamp;
-    let recordCount = 0;
-    for (const record of records) {
-      record.metadata = metadata;
-
-      this.$emit(record, {
-        ts: moment(record.createdTime).valueOf(),
-        summary: JSON.stringify(record.fields),
-        id: record.id,
-      });
-      if (!maxTimestamp || moment(record.createdTime).valueOf() > moment(maxTimestamp).valueOf()) {
-        maxTimestamp = record.createdTime;
-      }
-      recordCount++;
-    }
-    console.log(`Emitted ${recordCount} new records(s).`);
-    this._setLastTimestamp(maxTimestamp);
   },
 };