+++ components/basecamp/sources/new-event-by-webhook-type/new-event-by-webhook-type.mjs
@@ -4,17 +4,17 @@ import common from "../common/webhook.mjs";
 export default {
   ...common,
   key: "basecamp-new-event-by-webhook-type",
-  name: "New Event By Webhook Type (Instant)",
-  description: "Emit new event based on the selected webhook type. [See the docs here](https://github.com/basecamp/bc3-api/blob/master/sections/webhooks.md#create-a-webhook)",
-  version: "0.0.7",
+  name: "New Webhook Event (Instant)",
+  description: "Emit events of one or more selected types. [See the documentation](https://github.com/basecamp/bc3-api/blob/master/sections/webhooks.md#webhooks)",
+  version: "0.0.8",
   dedupe: "unique",
   type: "source",
   props: {
     ...common.props,
     webhookTypes: {
       type: "string[]",
-      label: "Webhook Types",
-      description: "Select types of webhook for listening.",
+      label: "Event Types",
+      description: "Select the types of events to be emitted. [See the documentation](https://github.com/basecamp/bc3-api/blob/master/sections/webhooks.md#webhooks) for more information.",
       options: constants.WEBHOOK_TYPE_OPTS,
     },
   },