+++ components/setmoreappointments/actions/create-appointment/create-appointment.mjs
@@ -5,7 +5,7 @@ export default {
   key: "setmoreappointments-create-appointment",
   name: "Create Appointment",
   description: "Create a new appointment in Setmore Appointments. [See the documentation](https://setmore.docs.apiary.io/#introduction/appointments/create-an-appointment)",
-  version: "0.0.1",
+  version: "0.0.2",
   type: "action",
   props: {
     setmore,