+++ components/zoho_crm/actions/create-object/create-object.mjs
@@ -1,11 +1,12 @@
 import common from "../common/common-objects.mjs";
+import { ConfigurationError } from "@pipedream/platform";
 
 export default {
   ...common,
   key: "zoho_crm-create-object",
   name: "Create Object",
   description: "Create a new object/module entry. [See the documentation](https://www.zoho.com/crm/developer/docs/api/v2/insert-records.html)",
-  version: "0.3.2",
+  version: "0.3.3",
   type: "action",
   async additionalProps() {
     const requiredProps = this.getRequiredProps(this.moduleType);
@@ -53,12 +54,18 @@ export default {
     });
     const objectData = {
       data: [
-        object,
+        this.parseFields(object),
       ],
     };
     const res = await zohoCrm.createObject(moduleType, objectData, $);
-    if (res.data[0].details.id) {
+
+    if (res.data[0].code === "SUCCESS") {
       $.export("$summary", `Successfully created new object with ID ${res.data[0].details.id}.`);
+    } else {
+      if (res.data[0].code === "INVALID_DATA") {
+        throw new ConfigurationError(`Error: Invalid data for field '${res.data[0].details.api_name}'. Expected data type: ${res.data[0].details.expected_data_type}`);
+      }
+      throw new ConfigurationError(res.data[0].message);
     }
     return res;
   },