+++ packages/connect-react/src/hooks/form-context.tsx
@@ -4,12 +4,15 @@ import {
 import isEqual from "lodash.isequal";
 import { useQuery } from "@tanstack/react-query";
 import type {
-  ComponentReloadPropsOpts, ConfigurableProp, ConfigurableProps, ConfiguredProps, V1Component, PropValue,
+  ComponentReloadPropsOpts, ConfigurableProp, ConfigurableProps, ConfiguredProps, V1Component,
 } from "@pipedream/sdk";
 import { useFrontendClient } from "./frontend-client-context";
 import type { ComponentFormProps } from "../components/ComponentForm";
 import type { FormFieldContext } from "./form-field-context";
-import { appPropError } from "./use-app";
+import {
+  appPropErrors, arrayPropErrors, booleanPropErrors, integerPropErrors,
+  stringPropErrors,
+} from "../utils/component";
 
 export type DynamicProps<T extends ConfigurableProps> = { id: string; configurableProps: T; }; // TODO
 
@@ -19,6 +22,7 @@ export type FormContext<T extends ConfigurableProps> = {
   configuredProps: ConfiguredProps<T>;
   dynamicProps?: DynamicProps<T>; // lots of calls require dynamicProps?.id, so need to expose
   dynamicPropsQueryIsFetching?: boolean;
+  errors: Record<string, string[]>;
   fields: Record<string, FormFieldContext<ConfigurableProp>>;
   id: string;
   isValid: boolean;
@@ -33,6 +37,13 @@ export type FormContext<T extends ConfigurableProps> = {
   submitting: boolean;
   userId: string;
 };
+
+export const skippablePropTypes = [
+  "$.service.db",
+  "$.interface.http",
+  "$.interface.apphook",
+  "$.interface.timer", // TODO add support for this (cron string and timers)
+]
 
 export const FormContext = createContext<FormContext<any /* XXX fix */> | undefined>(undefined); // eslint-disable-line @typescript-eslint/no-explicit-any
 
@@ -89,7 +100,7 @@ export const FormContextProvider = <T extends ConfigurableProps>({
   useEffect(() => {
     setEnabledOptionalProps({});
   }, [
-    component,
+    component.key,
   ]);
   // XXX pass this down? (in case we make it hash or set backed, but then also provide {add,remove} instead of set)
   const optionalPropIsEnabled = (prop: ConfigurableProp) => enabledOptionalProps[prop.name];
@@ -118,18 +129,24 @@ export const FormContextProvider = <T extends ConfigurableProps>({
     configuredProps,
     dynamicPropsId: dynamicProps?.id,
   };
+  const queryKeyInput = {
+    ...componentReloadPropsInput,
+  }
+
   const {
     isFetching: dynamicPropsQueryIsFetching,
     // TODO error
   } = useQuery({
     queryKey: [
       "dynamicProps",
+      queryKeyInput,
     ],
     queryFn: async () => {
       const { dynamicProps } = await client.componentReloadProps(componentReloadPropsInput);
       // XXX what about if null?
       // TODO observation errors, etc.
       if (dynamicProps) {
+        formProps.onUpdateDynamicProps?.(dynamicProps);
         setDynamicProps(dynamicProps);
       }
       setReloadPropIdx(undefined);
@@ -166,40 +183,41 @@ export const FormContextProvider = <T extends ConfigurableProps>({
 
   // these validations are necessary because they might override PropInput for number case for instance
   // so can't rely on that base control form validation
-  const propErrors = <T extends ConfigurableProps>(prop: ConfigurableProp, value: unknown): string[] => {
+  const propErrors = (prop: ConfigurableProp, value: unknown): string[] => {
     const errs: string[] = [];
-    if (value === undefined) {
-      if (!prop.optional) {
-        errs.push("required");
-      }
-    } else if (prop.type === "integer") { // XXX type should be "number"? we don't support floats otherwise...
-      if (typeof value !== "number") {
-        errs.push("not a number");
-      } else {
-        if (prop.min != null && value < prop.min) {
-          errs.push("number too small");
-        }
-        if (prop.max != null && value > prop.max) {
-          errs.push("number too big");
-        }
-      }
-    } else if (prop.type === "boolean") {
-      if (typeof value !== "boolean") {
-        errs.push("not a boolean");
-      }
-    } else if (prop.type === "string") {
-      if (typeof value !== "string") {
-        errs.push("not a string");
-      }
-    } else if (prop.type === "app") {
+    if (prop.optional || prop.hidden || prop.disabled || skippablePropTypes.includes(prop.type)) return []
+    if (prop.type === "app") {
       const field = fields[prop.name]
       if (field) {
         const app = field.extra.app
-        const err = appPropError({ value, app })
-        if (err) errs.push(err)
+        errs.push(...(appPropErrors({
+          prop,
+          value,
+          app,
+        }) ?? []))
       } else {
         errs.push("field not registered")
       }
+    } else if (prop.type === "boolean") {
+      errs.push(...(booleanPropErrors({
+        prop,
+        value,
+      }) ?? []))
+    } else if (prop.type === "integer") {
+      errs.push(...(integerPropErrors({
+        prop,
+        value,
+      }) ?? []))
+    } else if (prop.type === "string") {
+      errs.push(...(stringPropErrors({
+        prop,
+        value,
+      }) ?? []))
+    } else if (prop.type === "string[]") {
+      errs.push(...(arrayPropErrors({
+        prop,
+        value,
+      }) ?? []))
     }
     return errs;
   };
@@ -240,6 +258,9 @@ export const FormContextProvider = <T extends ConfigurableProps>({
     const newConfiguredProps: ConfiguredProps<T> = {};
     for (const prop of configurableProps) {
       if (prop.hidden) {
+        continue;
+      }
+      if (skippablePropTypes.includes(prop.type)) {
         continue;
       }
       // if prop.optional and not shown, we skip and do on un-collapse
@@ -322,6 +343,11 @@ export const FormContextProvider = <T extends ConfigurableProps>({
     const idx = configurableProps.findIndex((p) => p.name === prop.name);
     if (!enabled) {
       setConfiguredProp(idx, undefined);
+    } else if (__configuredProps?.[prop.name as keyof ConfiguredProps<T>] !== undefined) {
+      setConfiguredProp(
+        idx,
+        __configuredProps[prop.name as keyof ConfiguredProps<T>],
+      );
     } else if ("default" in prop && prop.default != null) {
       setConfiguredProp(idx, prop.default);
     }
@@ -331,7 +357,7 @@ export const FormContextProvider = <T extends ConfigurableProps>({
   const checkPropsNeedConfiguring = () => {
     const _propsNeedConfiguring = []
     for (const prop of configurableProps) {
-      if (!prop || prop.optional || prop.hidden) continue
+      if (!prop || prop.optional || prop.hidden || skippablePropTypes.includes(prop.type)) continue
       const value = configuredProps[prop.name as keyof ConfiguredProps<T>]
       const errors = propErrors(prop, value)
       if (errors.length) {
@@ -360,6 +386,7 @@ export const FormContextProvider = <T extends ConfigurableProps>({
     configuredProps,
     dynamicProps,
     dynamicPropsQueryIsFetching,
+    errors,
     fields,
     optionalPropIsEnabled,
     optionalPropSetEnabled,